moved the main method of TeamEvalResultTest to a new class in scripts
